add tests for help command
