+++ src/test/java/seedu/triplog/logic/commands/HelpCommandTest.java
@@ -94,4 +94,21 @@ public class HelpCommandTest {
         CommandResult result = new HelpCommand("  add  ").execute(model);
         assertTrue(result.getFeedbackToUser().startsWith("add "));
     }
+
+    @Test
+    public void addUsage_containsDateOptions() {
+        assertTrue(HelpCommand.ADD_USAGE.contains("sd/"));
+        assertTrue(HelpCommand.ADD_USAGE.contains("ed/"));
+    }
+
+    @Test
+    public void deleteUsage_containsIndexPlaceholder() {
+        assertTrue(HelpCommand.DELETE_USAGE.contains("<INDEX>"));
+    }
+
+    @Test
+    public void tagUsage_containsIndexAndTagNamePlaceholders() {
+        assertTrue(HelpCommand.TAG_USAGE.contains("<INDEX>"));
+        assertTrue(HelpCommand.TAG_USAGE.contains("<tag-name>"));
+    }
 }
