+++ src/test/java/seedu/triplog/ui/HelpWindowTest.java
@@ -13,7 +13,6 @@ import org.testfx.framework.junit5.Stop;
 
 import javafx.scene.input.KeyCode;
 import javafx.stage.Stage;
-import seedu.triplog.logic.commands.HelpCommand;
 
 @ExtendWith(ApplicationExtension.class)
 public class HelpWindowTest {
@@ -59,63 +58,6 @@ public class HelpWindowTest {
     public void isCloseKey_otherKey_returnsFalse() {
         assertFalse(HelpWindow.isCloseKey(KeyCode.A));
         assertFalse(HelpWindow.isCloseKey(KeyCode.ENTER));
-    }
-
-    @Test
-    public void isCloseKey_bKey_returnsFalse() {
-        assertFalse(HelpWindow.isCloseKey(KeyCode.B));
-    }
-
-    @Test
-    public void addUsage_containsCommandName() {
-        assertTrue(HelpCommand.ADD_USAGE.startsWith("add "));
-    }
-
-    @Test
-    public void editUsage_containsCommandName() {
-        assertTrue(HelpCommand.EDIT_USAGE.startsWith("edit "));
-    }
-
-    @Test
-    public void addUsage_containsDateOptions() {
-        assertTrue(HelpCommand.ADD_USAGE.contains("sd/"));
-        assertTrue(HelpCommand.ADD_USAGE.contains("ed/"));
-    }
-
-    @Test
-    public void deleteUsage_containsCommandName() {
-        assertTrue(HelpCommand.DELETE_USAGE.startsWith("delete "));
-    }
-
-    @Test
-    public void deleteUsage_containsIndexPlaceholder() {
-        assertTrue(HelpCommand.DELETE_USAGE.contains("<INDEX>"));
-    }
-
-    @Test
-    public void tagUsage_containsCommandName() {
-        assertTrue(HelpCommand.TAG_USAGE.startsWith("tag "));
-    }
-
-    @Test
-    public void findUsage_containsCommandName() {
-        assertTrue(HelpCommand.FIND_USAGE.startsWith("find "));
-    }
-
-    @Test
-    public void filterUsage_containsCommandName() {
-        assertTrue(HelpCommand.FILTER_USAGE.startsWith("filter "));
-    }
-
-    @Test
-    public void tagUsage_containsIndexAndTagNamePlaceholders() {
-        assertTrue(HelpCommand.TAG_USAGE.contains("<INDEX>"));
-        assertTrue(HelpCommand.TAG_USAGE.contains("<tag-name>"));
-    }
-
-    @Test
-    public void listUsage_containsCommandName() {
-        assertTrue(HelpCommand.LIST_USAGE.startsWith("list"));
     }
 
     @Test
